% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\docType{class}
\name{atlas-class}
\alias{atlas-class}
\title{Class atlas}
\description{
Class \code{atlas} This is the class used to create an object which will contains every dataframe and list created when creating submaps.
}
\section{Slots}{

\describe{
\item{\code{segments_list}}{the list of segments}

\item{\code{submaps_list}}{a list of submaps}

\item{\code{likelihood_summary}}{a dataframe with both likelihood0 and likelihood1 over the submaps}

\item{\code{estimation_summary}}{a dataframe with both a and f estimation over the submaps}

\item{\code{submap_summary}}{a dataframe with summary statitistics about the submaps}

\item{\code{HBD_recap}}{a dataframe with for one individual and for one marker a mean computation of all the HBD probabilities computed, on every individuals.}

\item{\code{FLOD_recap}}{a dataframe with for one individual and for one marker a mean computation of all the FLOD scores computed, on every individuals.}

\item{\code{HBDsegments}}{a list of dataframe with the HBDsegments, on all individuals.}

\item{\code{HFLOD}}{a dataframe with the value of HFLOD scores for every markers through all submaps.}

\item{\code{bedmatrix}}{a bed.matrix object (refer to gaston package)}

\item{\code{bySegments}}{a boolean indicating wheter the creation of summary statistics was made by segments (see documentation of Fantasio function)}

\item{\code{unit}}{the unit of the markers (cM or Bp).}

\item{\code{gap}}{the value of the gap used to pick marker when doing submaps by snps. (see function Fantasio for more infos)}

\item{\code{logisticRegression}}{the results of logistic regression}
}}

