% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSegmentsListByHotspots.r
\name{createSegmentsListByHotspots}
\alias{createSegmentsListByHotspots}
\title{Creation of a list of segments}
\usage{
createSegmentsListByHotspots(bedmatrix, intensity = 10,
  hotspot_version = "hg19", hotspot_file, verbose = TRUE,
  number_of_marker = 0)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{intensity}{Allows the user to select hotspots having recombination intensity higher than a threshold in cM/Mb (default is 10)}

\item{hotspot_version}{the version of the hotspots files (default is "hg19")}

\item{hotspot_file}{a personnal hotspots file with right columns in it ( see hotspots files )}

\item{verbose}{whether you want informations about the computation process (default is TRUE)}

\item{number_of_marker}{a threshold indicating the number of minimum marker in a segment (default is 50)}
}
\value{
an hotspots.segments object
}
\description{
This function is used to create a list of segments delimited by using hotspots files of the genome.
}
\details{
This function is used to create a list of chromosome list, which contains segments delimited by using the hotpost file.

The list is then wrapped under an object of class hotspots.segments for simplicity sake.

In each segment you have the index of all the makers in it, these index correpond to markers in the bedmatrix object.

The list structure can be analyse using `str()` function on the object (! careful the result can look messy if not handle properly)

You can give your own hotspots file, be sure to respect the format of the hotpots file, use a default hotpots file for it.
}
\examples{
 
#Please refer to vignette

}
\seealso{
read.bed.matrix

Fantasio
}
