% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubmapByHotspots.r
\name{createSubmapByHotpots}
\alias{createSubmapByHotpots}
\title{Creation of a submaps}
\usage{
createSubmapByHotpots(bedmatrix, segmentsList, epsilon = 0.001, fileName)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{segmentsList}{a list of segment for each chromosomes}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{fileName}{a fileName with the list of markers wanted for the submap}
}
\value{
return an hotspots.matrix object with some of his slots filled.
}
\description{
This function creates a submaps using the list segments created by using the hotspots in the genome
}
\details{
This function will iterates over the list of segments, then for each segments it will pick randomly one marker and put it into a vector.

Once the iteration over the list is over, the function will create an object and filled some of his slot.

If you are using a fileName, please make sure to have one marker per line
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListBySnps(bedMatrix)
submaps <- makeSubmapsBySnps(bedMatrix, 5, segmentList)

}
\seealso{
makeSubmapsByHotspots
}
