% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubmapByHotspots.r
\name{createSubmapByHotspots}
\alias{createSubmapByHotspots}
\title{Creation of a submaps}
\usage{
createSubmapByHotspots(bedmatrix, segmentsList, epsilon = 0.001,
  snpIndices)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{segmentsList}{a list of segment for each chromosomes}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{snpIndices}{(optional) You have the possibility to pass your own list of markers}
}
\value{
return an HostspotsMatrix object with some of his slots filled.
}
\description{
This function creates a submaps using the list segments created by using the hotspots in the genome
}
\details{
If snpIndices is given, the function creates a submap corresponding to the given SNPs.

Otherwise, this function iterates over the list of segments, then for each segments it picks randomly one marker.

This function is used internally in the package by the function makeAllSubmapsByHotspots
}
\examples{
 
#Please refer to vignette 


}
\seealso{
makeSubmapsByHotspots
}
