% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cytobands}
\alias{cytobands}
\alias{cytobands.b35}
\alias{cytobands.b36}
\alias{cytobands.b37}
\alias{cytobands.b38}
\title{Position of cytobands on the human chromosomes}
\format{A data frame with 32996 rows and 7 variables}
\usage{
data(cytobands.b35)
}
\description{
These datasets are data frames containing position of cytobands on human
chromosomes, for builds 35 (hg17), 36 (hg18), 37 (hg19) and 38 respectively.
The data were build starting from the data given in Bioconductor package 
\code{quantsmooth}.
}
\details{
The variables are:
\itemize{
  \item chr     Chromosome number.
  \item arm     Chromosome arm.
  \item band    Band name.
  \item stain   Band stain.
  \item beg.bp  Beginning position in base pairs.
  \item end.bp  Ending position in base pairs.
  \item beg.cM  Beginning position in centiMorgans.
  \item end.cM  Ending position in centiMorgans.
}
}
\keyword{datasets}
