% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_matrix.r
\docType{methods}
\name{dim,f.matrix-method}
\alias{dim,f.matrix-method}
\title{The generic dim method for f.matrix class}
\usage{
\S4method{dim}{f.matrix}(x)
}
\arguments{
\item{x}{a bed.matrix object}
}
\description{
The generic dim method for f.matrix class
}
