% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMatrix.r
\docType{methods}
\name{dim,fMatrix-method}
\alias{dim,fMatrix-method}
\title{The generic dim method for fMatrix class}
\usage{
\S4method{dim}{fMatrix}(x)
}
\arguments{
\item{x}{a bed.matrix object}
}
\description{
The generic dim method for fMatrix class
}
