% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_matrix.r
\docType{class}
\name{f.matrix-class}
\alias{f.matrix-class}
\title{Class f.matrix.}
\description{
Class defines a mother class for all the subclasses in the package Fantasio.
}
\section{Slots}{

\describe{
\item{\code{ncol}}{number of loci}

\item{\code{nrow}}{number of individual}

\item{\code{ped}}{the first 6 columns of a .ped}

\item{\code{map}}{id, chr, distance}

\item{\code{epsilon}}{value of epsilon = genotyping error rare, to use for computing log emission precalculated at initialisation}

\item{\code{delta.dist}}{difference of map$distance}

\item{\code{log.emiss}}{matrix of log proba emission if statut = 0 or 1(2 nb inds x nb msats)}

\item{\code{a}}{value of a estimated by festim}

\item{\code{f}}{value of f estimated by festim}

\item{\code{likelihood0}}{likelihood under H0 (f = 0)}

\item{\code{likelihood1}}{likelihood under H1}

\item{\code{p.lrt}}{likelihood ratio test}

\item{\code{HBD.prob}}{proba HBD = 1 ; one individual per column : dim = (nb inds x nb msats)}

\item{\code{FLOD}}{matrix of FLOD scores dim = (nb inds x nb msats)}
}}

