% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fantasio.r
\name{Fantasio}
\alias{Fantasio}
\title{Creation of submaps using gap between markers}
\usage{
Fantasio(bedmatrix, segments = c("Hotspots", "Distance"), segment.options,
  n = 100, n.cores = 1, epsilon = 0.001, run.festim = TRUE, list.id,
  run.proba = TRUE, recap.by.segments = FALSE, verbose = TRUE,
  debug = FALSE, threshold = 0.5, q = 1e-04, quality = 95,
  n.consecutive.marker = 5)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{segments}{The method to use for submap creation ("Hotspots" or "Distance")}

\item{n}{the number of submaps wanted}

\item{n.cores}{the number of cores to use if you want to compute submaps using parellelism}

\item{epsilon}{the value of epsilon for submaps}

\item{run.festim}{whether you want to computes a, f, p.lrt, likelihood0/1 for each submaps}

\item{list.id}{a list of individuals from which to computes HBD, 
FLOD score and HFLOD score (see details for more informations)}

\item{run.proba}{whether you want to computes HBD, FLOD score and HFLOD score}

\item{recap.by.segments}{if you want the summary of probabilities by snps or by segments}

\item{verbose}{whether you want informations about computations}

\item{debug}{whether you want advanced output for the computation process}

\item{threshold}{the value of the threshold when finding HBD segment}

\item{q}{Allows the user to choose the assumed frequency of the mutation
involved in the disease for each individual. Default is 0.0001.}

\item{quality}{Allows the user to choose the minimal quality 
to include an inbred individual into the analysis. Default is 95.}

\item{n.consecutive.marker}{the number of consecutive marker with a 
probabilities equal or greater to the value of threshold, to be use to fing HBD segments}
}
\description{
Creates n submaps and generates summaries following the specified options
}
\details{
manuel a finir !!!!!
}
