% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/festim.r
\name{festim}
\alias{festim}
\title{Computation of a and f}
\usage{
festim(x, n.cores = 1, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{x}{a submap matrix}

\item{verbose}{whether you want informations about the computation process (default is TRUE)}

\item{debug}{whether you want advanced output about the computation process (default is FALSE)}
}
\value{
returns the submap matrix with the slot a, f, p.lrt, likelihood0 and likelihood1 filled.
}
\description{
This function is used to compute a and f and other statistical values (p.lrt, likelihood0, likelihood1)
}
\details{
This function is using the optim function with an L-BGFS method to find the best value for a and f
}
\examples{
 
#Please refer to vignette 


}
