% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQQPlot.r
\name{ggQQPlot}
\alias{ggQQPlot}
\title{Fonction to plot QQplot with confidence interval 
Marie-Sophie Ogloblinsky}
\usage{
ggQQPlot(ps, ci = 0.95)
}
\arguments{
\item{ps}{a vector of p_value}

\item{ci}{the confidence interval (95% default)}
}
\description{
Fonction to plot QQplot with confidence interval 
Marie-Sophie Ogloblinsky
}
