% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmHBD.r
\name{glmHBD}
\alias{glmHBD}
\title{Logistic regression on HBD probability or FLOD score}
\usage{
glmHBD(x, expl_var, covar_df, covar, n.cores = 1, run = FALSE, phen.code)
}
\arguments{
\item{x}{an atlas object}

\item{expl_var}{the explanatory variable 'FLOD' or 'HBD_prob'}

\item{covar_df}{a dataframe containing covariates}

\item{covar}{covariates of interest such as 'age', 'sex' , ...
if missing, all covariates of the dataframe are considered}

\item{n.cores}{number of cores for parallelization calculation (default = 1)}

\item{run}{whether the fonction is called or not (default = FALSE)}

\item{phen.code}{phenotype coding :
       - 'R' : 0:control ; 1:case ; NA:unknown (default)
       - 'plink' : 1:control ; 2:case ; 0/-9/NA:unknown
if 'plink' the function automatically convert it to 'R' to run logistic regression}
}
\description{
Logistic regression on HBD probability or FLOD score
}
