% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBDPlotChr.r
\name{HBDPlotChr}
\alias{HBDPlotChr}
\title{plot of HBD segment}
\usage{
HBDPlotChr(
  Submaps,
  ROH,
  unit = "cM",
  chr,
  list.ids,
  regions,
  outfile,
  inbred = FALSE,
  build = 37
)
}
\arguments{
\item{Submaps}{a list.submap object}

\item{ROH}{a data frame from which the segments will be plotted (optional)}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{chr}{the chromosome number from which to plot HBD segment}

\item{list.ids}{a vector containing a list of individuals from which only the HBDsegments for this chromosome will be ploted (optional)}

\item{regions}{a specific region to be enlighted in the plot (optional)}

\item{outfile}{a name for the plot (optional)}

\item{inbred}{whether you want to plot only inbred individuals or not (default is FALSE)}

\item{build}{the value of the build to use to plot chromosome in the plot value accepted are 35, 36, 37, 38 (default is 37)}
}
\value{
return a plot of the chromosome HBDsegments for all the individual
}
\description{
This function plots the HBDsegments for a given chromosome and all the individuals
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
\examples{
 
#Please refer to vignette 


}
\seealso{
Fantasio

setHFLOD
}
