% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBDsegments.r
\name{HBDsegments}
\alias{HBDsegments}
\title{Find HBDsegments}
\usage{
HBDsegments(submaps, n.consecutive.marker = 5, threshold = 0.5,
  recap.by.segments = FALSE)
}
\arguments{
\item{submaps}{a submapsList object}

\item{n.consecutive.marker}{the number of consecutive marker with a probabilitie equal or greater to the value of the threshold, to be use to fing HBDsegments (default is 5)}

\item{threshold}{the minimum value of HBD probabilities given for a marker (default is 0.5)}

\item{recap.by.segments}{if the summary files has to be computed considering segments or snps (defaut is FALSE)}
}
\value{
This function returns a list of dataframe with 11 columns :

- individual

- family

- status

- start

- end

- size

- chromosome

- start_pos

- end_pos

- start_dist

- end_dist
}
\description{
This function creates HBDsegments used for creating plots
}
\details{
The threshold is the minimum value from which we consider the marker is HBD. From this marker we want a minumum number of consecutive markers to create a HBD segment.
(argument : n.consecutive.marker)
}
\seealso{
setSummary

setHBDprob
}
