% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFLODplotChr.r
\name{HFLODplotChr}
\alias{HFLODplotChr}
\title{Plot of the HFLOD}
\usage{
HFLODplotChr(submaps, unit = "cM", chr, regions, color2 = "green4",
  MA = TRUE, nbSNP_MA = 50)
}
\arguments{
\item{submaps}{a submapsList object}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{chr}{the chromosome number from which to plot HFLOD score}

\item{regions}{a matrix containing the value to ve highlighted in the plot}

\item{color2}{the color of the regions highlighted (default is "green4")}

\item{MA}{a boolean indicating whether a red line has to be drawn for the moving average}

\item{nbSNP_MA}{number of SNP for the moving average (default is 50)}
}
\value{
This function returns a manhattan plot of all the HFLOD score over all the chromosome
}
\description{
This fonction plot the HFLOD score for a chromosome
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

- the chromosome number

- start

- end
}
\examples{
 
#Please refer to vignette 

}
\seealso{
setHFLOD
}
