% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotsSegments.r
\docType{class}
\name{HostspotsSegments-class}
\alias{HostspotsSegments-class}
\title{Class HostspotsSegments}
\description{
Class \code{HostspotsSegments} This is the class that encapsulate the list of segments created by using the hotspots in the genome.
}
