% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.segments.r
\docType{class}
\name{hotspot.segments-class}
\alias{hotspot.segments-class}
\title{Class hotspot.segments}
\description{
Class \code{hotspot.segments} This is the class that encapsulate the list of segments created by using the hotspots in the genome.
}
