% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hotspot_hg18}
\alias{hotspot_hg18}
\title{List of hotspots throughout the genome version 18}
\format{A data frame with 32990 rows and 7 variables}
\usage{
data(hotspot_hg18)
}
\description{
A dataset containing information about hotspots in the genome.  The variables are as follows:
}
\details{
\itemize{
  \item Chromosome. the number of the chromosome which contains the hotspots
  \item Centre. The center of the hotspots in Mb
  \item Start. Start of the hotspot
  \item End.  End of the hotspot
  \item Widthkb. the width of the hotposts in kb
  \item IntensitycMMb. the intensity
  \item TotaldistancecM. the total distance in cM
}
}
\keyword{datasets}
