% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{initialize,atlas-method}
\alias{initialize,atlas-method}
\title{Constructor method of atlas.}
\usage{
\S4method{initialize}{atlas}(
  .Object,
  submaps,
  bedmatrix,
  segments_list,
  bySegments,
  unit = NULL,
  gap = NULL
)
}
\arguments{
\item{.Object}{the object type}

\item{submaps}{a list of submaps}

\item{bedmatrix}{a bed.matrix object}

\item{segments_list}{a list of segments object}

\item{bySegments}{a boolean}

\item{unit}{the unit in which are markers (cM or Bp)}

\item{gap}{the gap used to create segments in the By Distance method}

\item{logReg}{list of logistic regression results on adjusted and unadjusted data}
}
\description{
Constructor method of atlas.
}
