% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMatrix.r
\name{initialize,fMatrix-method}
\alias{initialize,fMatrix-method}
\title{Constructor method of fMatrix}
\usage{
\S4method{initialize}{fMatrix}(.Object, ncol, nrow, ped, map, log.emiss, epsilon = 0.001)
}
\arguments{
\item{.Object}{the object type}

\item{ncol}{number of loci}

\item{nrow}{number of individual}

\item{ped}{the first 6 columns of a .ped}

\item{map}{id, chr, distance}

\item{log.emiss}{matrix of log proba emission if statut = 0 or 1(2 nb inds x nb msats)}

\item{epsilon}{value of epsilon = genotyping error rare, to use for computing log emission precalculated at initialisation}
}
\description{
Constructor method of fMatrix
}
