% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotsMatrix.r
\docType{methods}
\name{initialize,HostspotsMatrix-method}
\alias{initialize,HostspotsMatrix-method}
\title{Constructor method of HostspotsMatrix.}
\usage{
\S4method{initialize}{HostspotsMatrix}(.Object, ncol, nrow, submap, ped,
  map, log.emiss, epsilon = 0.001)
}
\arguments{
\item{.Object}{the object type}

\item{ncol}{number of loci}

\item{nrow}{number of individual}

\item{submap}{a list of submaps}

\item{ped}{the first 6 columns of a .ped}

\item{map}{id, chr, distance}

\item{log.emiss}{matrix of log proba emission if statut = 0 or 1(2 nb inds x nb msats)}

\item{epsilon}{value of epsilon = genotyping error rare, to use for computing log emission precalculated at initialisation}
}
\description{
Constructor method of HostspotsMatrix.
}
