% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.submaps.r
\docType{methods}
\name{initialize,list.submaps-method}
\alias{initialize,list.submaps-method}
\title{Constructor method of list.submaps.}
\usage{
\S4method{initialize}{list.submaps}(.Object, submaps, bedmatrix, segments_list,
  bySegments, unit = NULL, gap = NULL)
}
\arguments{
\item{.Object}{the object type}

\item{submaps}{a list of submaps}

\item{bedmatrix}{a bed.matrix object}

\item{segments_list}{a list of segments object}

\item{bySegments}{a boolean}

\item{unit}{the unit in which are markers (cM or Bp)}

\item{gap}{the gap used to create segments in the By Distance method}
}
\description{
Constructor method of list.submaps.
}
