% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.segments.r
\docType{methods}
\name{initialize,snps.segments-method}
\alias{initialize,snps.segments-method}
\title{Constructor method of snps.segments.}
\usage{
\S4method{initialize}{snps.segments}(.Object, gap, unit, segments)
}
\arguments{
\item{.Object}{the type of the object ( here snps.segments)}

\item{gap}{the gap used to creates the segments}

\item{unit}{the unit in which are the markers}

\item{segments}{the list of segments}
}
\description{
Constructor method of snps.segments.
}
