% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsSegments.r
\name{initialize,snpsSegments-method}
\alias{initialize,snpsSegments-method}
\title{Constructor method of snpsSegments.}
\usage{
\S4method{initialize}{snpsSegments}(.Object, gap, unit, segments)
}
\arguments{
\item{.Object}{the type of the object ( here snpsSegments)}

\item{gap}{the gap used to creates the segments}

\item{unit}{the unit in which are the markers}

\item{segments}{the list of segments}
}
\description{
Constructor method of snpsSegments.
}
