% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAllSubmapsByHotspots.r
\name{makeAllSubmapsByHotspots}
\alias{makeAllSubmapsByHotspots}
\title{Creation of submaps using hotposts in the genome}
\usage{
makeAllSubmapsByHotspots(bedmatrix, n = 100,
  segmentsList = createSegmentsListByHotspots(bedmatrix), n.cores = 1,
  epsilon = 0.001, run.festim = TRUE, list.id, run.proba = TRUE,
  recap.by.segments = FALSE, verbose = TRUE, debug = FALSE,
  threshold = 0.5, q = 1e-04, quality = 95, n.consecutive.marker = 5)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{n}{the number of submaps wanted}

\item{segmentsList}{a list of segment for each chromosomes}

\item{n.cores}{the number of cores to use if you want to compute submaps using parellelism}

\item{epsilon}{the value of epsilon for submaps}

\item{run.festim}{whether you want to computes a, f, p.lrt, likelihood0/1 for each submaps}

\item{list.id}{a list of individuals from which to computes HBD, FLOD score and HFLOD score (see details for more informations)}

\item{run.proba}{whether you want to computes HBD, FLOD score and HFLOD score}

\item{recap.by.segments}{if you want the summary of probabilities by snps or by segments}

\item{verbose}{whether you want informations about computations}

\item{debug}{whether you want advanced output for the computation process}

\item{threshold}{the value of the threshold when finding HBD segment}

\item{q}{Allows the user to choose the assumed frequency of 
the mutation involved in the disease for each individual. Default is 0.0001.}

\item{quality}{Allows the user to choose the minimal quality  to include an 
inbred individual into the analysis. Default is 95 .}

\item{n.consecutive.marker}{the number of consecutive marker with a probabilitie
equal or greater to the value of threshold, to be use to fing HBD segments}
}
\value{
return a new list object containing every dataframe and object created
}
\description{
This function creates N submaps and allows the creation of summary files
}
\details{
This function is used to create submaps by randomly picking a marker
 in each segment in the list of segments given by the argument segmentsList.

After the creation of one submap is finished, the function pass it 
to `festim`, a function that will computes different values (a, f, p.lrt, likelihood0, likelihood1)

After that we then creates the different summary to interpret easily 
the results of the computation.

This is done by the function `setSummary`. This function will call all 
of the function neccessary to obtained the different summary needed for our data.

You can check all the summary after they are been created by accesing their slots.

When using recap.by.segments with true value, we then consider that the 
snps picked randomly in a segment is a representant of that segment.

you can pass several arguments to the list.id arguments : 
\itemize{
 \item{ignore this parameter if you want to compute HBD, FLOD and HFLOD for INRED individuals and with a QUALITY}
 \item{enter a list of individual for a computation of HBD, FLOD 
 score HFLOD score for them, each element of the vector should contains the familyId_individualId}
 \item{the character "all" for a computation of HBD, FLOD score and HFLOD score for every individual}
}

When doing a number N of submaps the different values computed from the 
snps picked randomly (a, f, p.lrt, ...) in a marker is considered different values 
for the same segments.
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListByHotspots(bedMatrix)
submaps <- makeSubmapsByHotspots(bedMatrix, 5, segmentList)
}
\seealso{
\code{\link{makeSubmapsBySnps}}

\code{\link{createSegmentsListByHotspots}}

\code{\link{festim}}

\code{\link{set.HBD.prob}}

\code{\link{set.FLOD}}

\code{\link{set.HFLOD}}

\code{\link{recap}}

\code{\link{setSummary}}

\code{\link{submapLikelihood}}

\code{\link{submapEstim}}

\code{\link{summaryMarker}}

\code{\link{submapSummary}}

\code{\link{HBD.segments}}
}
