% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAtlasByHotspots.r
\name{makeAtlasByHotspots}
\alias{makeAtlasByHotspots}
\title{Creation of submaps based on hotposts in the genome}
\usage{
makeAtlasByHotspots(
  bedmatrix,
  n = 100,
  segmentsList = segmentsListByHotspots(bedmatrix),
  n.cores = 1,
  epsilon = 0.001
)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{n}{the number of submaps wanted(default is 100)}

\item{segmentsList}{a list of segment for each chromosomes}

\item{n.cores}{the number of cores to use if you want to compute submaps using parellelism (default is 1)}

\item{epsilon}{genotype error rate (default is 0.001)}
}
\value{
return a new list object containing every dataframe and object created
}
\description{
This function creates N submaps and allows the creation of summary files
}
\details{
This function is used to create submaps by randomly picking a marker
 in each segment in the list of segments given by the argument segmentsList.

After the creation of one submap is finished, the function pass it 
to `festim`, a function that will computes different values (a, f, p.lrt, likelihood0, likelihood1)

After that we then creates the different summary to interpret easily 
the results of the computation.

This is done by the function `setSummary`. This function will call all 
of the function neccessary to obtained the different summary needed for our data.

You can check all the summary after they are been created by accesing their slots.

When using recap.by.segments with true value, we then consider that the 
snps picked randomly in a segment is a representant of that segment.

When doing a number N of submaps using the options recap.by.segments,

the different values computed from the

snps picked randomly (a, f, p.lrt, ...) in a segments is considered different values

for the same segments.
}
\examples{
 
#Please refer to vignette 

}
\seealso{
Fantasio

makeAtlasByDistance

segmentsListByHotspots

festim

setHBDProb

setFLOD

setHFLOD

recap

setSummary

submapLikelihood

submapEstim

submapSummary

HBDSegments
}
