% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSubmapByDistance.r
\name{makeSubmapByDistance}
\alias{makeSubmapByDistance}
\title{Creation of a submaps}
\usage{
makeSubmapByDistance(bedmatrix, segmentsList, epsilon = 0.001, snpIndices)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{segmentsList}{a list of segments for each chromosome}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{snpIndices}{(optional) You have the possibility to pass your own list of markers}
}
\value{
return an snpsMatrix object
}
\description{
This function creates a submap using the list segments created by using gap between markers
}
\details{
If snpIndices is given, the function creates a submap corresponding to the given SNPs.

Otherwise, the function iterates over the list of segments, then for each segments it picks randomly one marker 
and from this marker choses a marker every "step" step from back to end.
}
\examples{
 
#Please refer to vignette 


}
\seealso{
makeAtlasByDistance
}
\keyword{internal}
