% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerRepresentation.r
\name{markerRepresentation}
\alias{markerRepresentation}
\title{Markers selected through the submaps}
\usage{
markerRepresentation(atlas)
}
\arguments{
\item{atlas}{an atlas}
}
\value{
A data frame with columns: 
\describe{
 \item{id}{id of the snps}
 \item{chr}{chromosome}
 \item{pos}{position in bp}
 \item{dist}{position in cM}
 \item{Freq}{number of times the marker has been picked}
}
}
\description{
This function gives for each marker (picked at least once) 
the number of times it has been selected in a submap.
}
\seealso{
setSummary
}
