% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerSummary.r
\name{markerSummary}
\alias{markerSummary}
\title{Number of marker selected in a submap}
\usage{
markerSummary(submaps)
}
\arguments{
\item{submaps}{a list.submaps object}
}
\value{
this function returns a dataframe.
}
\description{
This function gives for each submap the number of marker selected.
}
\details{
the dataframe contains the following elements :

- by columns is the number of markers used

- by row is each submap
}
\seealso{
setSummary
}
