% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoConverter.r
\name{phenoConverter}
\alias{phenoConverter}
\title{Fonction to convert phenotype when the phenotype coded with plink
convert plink to R coding phenotype}
\usage{
phenoConverter(x, phen.code = c("R", "plink"))
}
\arguments{
\item{x}{the bedmatrix}

\item{phen.code}{type of coding phenotype 'R' or 'plink'
R : 0:control ; 1:cases ; NA:unknown (default)
plink : 1:control ; 2:cases ; 0/-9/NA:unknown}
}
\value{
the bed matrix with converted phenotype
}
\description{
Fonction to convert phenotype when the phenotype coded with plink
convert plink to R coding phenotype
}
