% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ROH.segments.chr.r
\name{plot.ROH.segments.chr}
\alias{plot.ROH.segments.chr}
\title{Creation of plot using ROH file}
\usage{
\method{plot}{ROH.segments.chr}(ROHfile, submaps, unit = "cM", chr,
  outfile, listid, regions, build)
}
\arguments{
\item{ROHfile}{an ROH file}

\item{submaps}{a atlas object}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{chr}{the chromosome number from which to plot ROH}

\item{outfile}{the name of the plot}

\item{listid}{a vector containing the family id follow by the individual id}

\item{regions}{a matrix containing the value to ve highlighted in the plot}

\item{build}{the value of the build to use to plot chromosome in the plot value accepted are 35, 36, 37, 38 (default is 37)}
}
\value{
This function returns a plot of the HBDsegments for a specific chromosome
}
\description{
This function is use to plot HBDsegments for a chromosome using an ROH file
}
\details{
Use this function when you want to use an ROH file to plot your HBDsegments for a specific chromosome.

Two unit are accepted : "Bases" or "cM".

If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
