% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ROH.segments.id.r
\name{plot.ROH.segments.id}
\alias{plot.ROH.segments.id}
\title{Creation of plot using ROH file}
\usage{
\method{plot}{ROH.segments.id}(Submaps, ROHfile, unit = "cM", regions,
  outfile, family.id, individual.id, build)
}
\arguments{
\item{ROHfile}{an ROH file}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{regions}{a matrix containing the value to ve highlighted in the plot}

\item{outfile}{the name of the plot}

\item{build}{the value of the build to use to plot chromosome in the plot value accepted are 35, 36, 37, 38 (default is 37)}

\item{submaps}{a list.submaps object}

\item{family_id}{the individual id of the individual wanted}

\item{individual_id}{the family id of the individual wanted}

\item{save_file}{whether you want or not to save the plot}
}
\value{
This function returns a plot of the HBD segments for a specific individual
}
\description{
This function is use to plot HBD segments for a specific infdividual using an ROH file
}
\details{
Use this function when you want to use an ROH file to plot your HBD segments for a specific individual.

Two unit are accepted : "Bases" or "cM".

If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
