% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROHSegmentsId.r
\name{plotROHSegmentsId}
\alias{plotROHSegmentsId}
\title{Creation of plot using ROH file}
\usage{
plotROHSegmentsId(
  Submaps,
  ROH,
  unit = "cM",
  regions,
  outfile,
  famid,
  id,
  build
)
}
\arguments{
\item{Submaps}{a atlas object}

\item{ROH}{a data frame}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{regions}{a matrix containing the value to ve highlighted in the plot}

\item{outfile}{the name of the plot}

\item{famid}{the family id of the individual wanted}

\item{id}{the individual id of the individual wanted}

\item{build}{the value of the build to use to plot chromosome in the plot value accepted are 35, 36, 37, 38 (default is 37)}
}
\value{
This function returns a plot of the HBDsegments for a specific individual
}
\description{
This function plots HBDsegments for a specific individual using an ROH file
}
\details{
Use this function when you want to use an ROH file to plot your HBDsegments for a specific individual.

Two unit are accepted : "Bases" or "cM".

If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
\keyword{internal}
