% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.r
\name{recap}
\alias{recap}
\title{HBD/FLOD recap}
\usage{
recap(atlas, recap.by.segments = atlas@bySegments)
}
\arguments{
\item{atlas}{a atlas object}

\item{recap.by.segments}{: whether you want the recap by segments or snps}
}
\value{
This function returns a list of dataframes.
}
\description{
This function creates HBD and FLOD recap dataframe.
}
\details{
For each individual and each marker the function computes

the mean value of every HBD probabilities computed through the submaps.

The same is done with the FLOD scores.

This function returns a list of two dataframes with HBD and FLOD in it.
}
\seealso{
Fantasio

makeAtlasByDistance

segmentsListByHotspots

festim

setHBDProb

setFLOD

setHFLOD

HBDSegments
}
