% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsListByDistance.r
\name{segmentsListByDistance}
\alias{segmentsListByDistance}
\title{Creation of a list of segments}
\usage{
segmentsListByDistance(
  bedmatrix,
  gap = 0.5,
  minMarkers = 50,
  nbSegments = 20,
  unit = "cM",
  verbose = TRUE
)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{gap}{the minimum gap after which a segment is created}

\item{minMarkers}{the minimum number of markers in a mini-segment}

\item{nbSegments}{the number of mini-segments which the segment will be split}

\item{unit}{the \code{gap} unit, "cM" (centirmorgans) or "Mb" (million of bases)}

\item{verbose}{if \code{TRUE}, displays information on the process}
}
\value{
an snpsSegments object
}
\description{
Creates a list of segments delimited by gaps in the genome.
}
\details{
This function is used to create an object of class snpsSegments. It contains a list of segments 
delimited by gaps between available markers. Each segment is split in \code{nbSegments} mini-segments.

A segment is not split if it would result in mini-segments shorter that \code{minMarkers}.
The indices in the mini-segments correspond to SNPs in \code{bedmatrix}.
}
\examples{
 
#Please refer to vignette 

}
\seealso{
\code{\link{Fantasio}}, \code{\link{segmentsListByDistance}}
}
