% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsListByHotspots.r
\name{segmentsListByHotspots}
\alias{segmentsListByHotspots}
\title{Creation of a list of segments}
\usage{
segmentsListByHotspots(bedmatrix, intensity = 10,
  hotspots = hotspot_hg19, minMarkers = 0, verbose = TRUE)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{intensity}{hotspots intensity threshold in cM/Mb}

\item{hotspots}{a data frame of recombination rates}

\item{minMarkers}{minimum number of markers in a segment}

\item{verbose}{if \code{TRUE}, displays information on the process}
}
\value{
an hotspots.segments object
}
\description{
Creates a list of segments delimited by recombination hotspots
}
\details{
This function creates an object of class hotspots.segments, containing a list of segments delimited
by hotspots. The object is a list of list of vectors indices of SNPs. There are as many sublists as
chromosomes. The indices correspond to SNPs in \code{bedmatrix}.

The user can provide a hotspots data frame with a format similar to hotspot_hg19.
}
\examples{
 
#Please refer to vignette

}
\seealso{
\code{\link{Fantasio}}, \code{\link{segmentsListByDistance}}
}
