% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsListSummary.r
\name{segmentsListSummary}
\alias{segmentsListSummary}
\title{Summary of marker picked}
\usage{
segmentsListSummary(segmentList)
}
\arguments{
\item{segmentList}{A list of segments}
}
\description{
This function returns the number of segments and number of marker in the list of segments outputed by the function 
`segmentsListByHotspots`
}
\examples{
#Please refer to vignette 


}
