% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.p.r
\name{set.allelic.frequencies}
\alias{set.allelic.frequencies}
\title{Set p value in the bed.matrix}
\usage{
set.allelic.frequencies(x, freq)
}
\arguments{
\item{x}{a bed.matrix}

\item{freq}{vector of allelic frequencies}
}
\description{
Use this function if you want to change the allelic frequencies in the bed.matrix object.
}
