% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.status.r
\name{set.status}
\alias{set.status}
\title{Set status in the bed.matrix}
\usage{
set.status(bedmatrix, status = 2)
}
\arguments{
\item{bedmatrix}{a bed.matrix}

\item{status}{value accepted are 0, -9, 1, 2 (default is 2)}
}
\description{
Use this function if you want to change the status of your individuals in the bed.matrix
}
