% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAlleleFrequencies.r
\name{setAlleleFrequencies}
\alias{setAlleleFrequencies}
\title{This function assign allele frequencies from a reference data set
Use this function if you want to change the allelic frequencies in the bed.matrix object.}
\usage{
setAlleleFrequencies(x, x_ref)
}
\arguments{
\item{x}{the bed.matrix to study}

\item{x_ref}{the reference bed.matrix to set allelic frequencies}
}
\description{
This function assign allele frequencies from a reference data set
Use this function if you want to change the allelic frequencies in the bed.matrix object.
}
