% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAllelicFrequencies.r
\name{setAllelicFrequencies}
\alias{setAllelicFrequencies}
\title{Set p value in the bed.matrix}
\usage{
setAllelicFrequencies(x, freq)
}
\arguments{
\item{x}{a bed.matrix}

\item{freq}{vector of allelic frequencies}
}
\description{
Use this function if you want to change the allelic frequencies in the bed.matrix object.
}
