% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHBDProb.r
\name{setHBDProb}
\alias{setHBDProb}
\title{Computation of HBD probabilities}
\usage{
setHBDProb(atlas, w.id)
}
\arguments{
\item{atlas}{A atlas object}

\item{w.id}{The indices of individals for which HBD probabilities are computed}
}
\value{
the atlas object with each HBD.prob slot of each submaps in the slot submaps_list computed
}
\description{
This function is used to compute HBD probabilities on individuals in a sample
}
\details{
This function iterates over the slots submaps_list of the atlas object.

For each submaps in the slots submaps_list of the object, the slot HBD.prob will be filled with a matrix of dimension : number_individual x number_of_markers
}
\seealso{
setFLOD

setHFLOD
}
