% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHBDProbAndFLOD.r
\name{setHBDProbAndFLOD}
\alias{setHBDProbAndFLOD}
\title{Computation of HBD probabilities and FLOD scores by segments}
\usage{
setHBDProbAndFLOD(atlas, w.id, q = 1e-04)
}
\arguments{
\item{atlas}{an atlas object}

\item{w.id}{The indices of individals for which HBD probabilities are computed}
}
\value{
the atlas object with HBD_recap and FLOD_recap by segments
}
\description{
This function is used to compute HBD probabilities and FLOD scores on individuals in a sample
}
\details{
This function iterates over the slots submaps_list of the atlas object.

For each submaps in the slots submaps_list of the object, the slots HBD_recap and FLOD_recap will be updated with a matrix of dimension : number_individual x number_of_segments
}
\seealso{
\code{\link{setHBDProbAndFLODBySnps}}

\code{\link{setHFLOD}}
}
