% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHBDProbAndFLODBySnps.r
\name{setHBDProbAndFLODBySnps}
\alias{setHBDProbAndFLODBySnps}
\title{Computation of HBD probabilities and FLOD scores by snps}
\usage{
setHBDProbAndFLODBySnps(atlas, w.id, q = 1e-04)
}
\arguments{
\item{atlas}{an atlas object}

\item{w.id}{The indices of individals for which HBD probabilities are computed}
}
\value{
the atlas object with HBD_recap and FLOD_recap by snps
}
\description{
This function is used to compute HBD probabilities and FLOD scores on individuals in a sample
}
\details{
This function iterates over the slots submaps_list of the atlas object.

For each submaps in the slots submaps_list of the object, the slots HBD_recap and FLOD_recap will be updated with a matrix of dimension : number_individual x number_of_markers
}
\seealso{
\code{\link{setHFLOD}}
}
