% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHFLOD.r
\name{setHFLOD}
\alias{setHFLOD}
\title{Computation of HFLOD scores}
\usage{
setHFLOD(atlas, w.id)
}
\arguments{
\item{atlas}{a atlas object}

\item{w.id}{vector of indices to be considered in the FLOD_recap matrix}
}
\value{
the atlas object with it's slot HFLOD filled with a matrix of dimensions : number_of_marker x 2. The first column is the value of HFLOD for the marker. 
The second value is the moving average probability.
}
\description{
This function is used to compute HFLOD scores on individuals in a sample
}
\details{
This function iterates over the slots submaps_list of the atlas object.
}
\seealso{
setHBDProb

setFLOD
}
