% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\docType{methods}
\name{show,atlas-method}
\alias{show,atlas-method}
\title{Show method of atlas.}
\usage{
\S4method{show}{atlas}(object)
}
\arguments{
\item{object}{an atlas object}
}
\description{
Show method of atlas.
}
