% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_matrix.r
\docType{methods}
\name{show,f.matrix-method}
\alias{show,f.matrix-method}
\title{show method of f.matrix class}
\usage{
\S4method{show}{f.matrix}(object)
}
\arguments{
\item{object}{an f.matrix object}
}
\description{
show method of f.matrix class
}
