% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMatrix.r
\name{show,fMatrix-method}
\alias{show,fMatrix-method}
\title{show method of fMatrix class}
\usage{
\S4method{show}{fMatrix}(object)
}
\arguments{
\item{object}{an fMatrix object}
}
\description{
show method of fMatrix class
}
