% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotsMatrix.r
\docType{methods}
\name{show,HostspotsMatrix-method}
\alias{show,HostspotsMatrix-method}
\title{Show method of HostspotsMatrix.}
\usage{
\S4method{show}{HostspotsMatrix}(object)
}
\arguments{
\item{object}{an HostspotsMatrix object}
}
\description{
Show method of HostspotsMatrix.
}
