% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotsSegments.r
\name{show,HostspotsSegments-method}
\alias{show,HostspotsSegments-method}
\title{Show method of HostspotsSegments.}
\usage{
\S4method{show}{HostspotsSegments}(object)
}
\arguments{
\item{object}{the HostspotsSegments}
}
\description{
Show method of HostspotsSegments.
}
