% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.segments.r
\docType{methods}
\name{show,hotspot.segments-method}
\alias{show,hotspot.segments-method}
\title{Show method of hotspot.segments.}
\usage{
\S4method{show}{hotspot.segments}(object)
}
\arguments{
\item{object}{the hotspot.segments}
}
\description{
Show method of hotspot.segments.
}
