% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspots.matrix.r
\docType{methods}
\name{show,hotspots.matrix-method}
\alias{show,hotspots.matrix-method}
\title{Show method of hotspots.matrix.}
\usage{
\S4method{show}{hotspots.matrix}(object)
}
\arguments{
\item{object}{an hotspots.matrix object}
}
\description{
Show method of hotspots.matrix.
}
