% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.submaps.r
\docType{methods}
\name{show,list.submaps-method}
\alias{show,list.submaps-method}
\title{Show method of list.submaps.}
\usage{
\S4method{show}{list.submaps}(object)
}
\arguments{
\item{object}{an list.submaps object}
}
\description{
Show method of list.submaps.
}
