% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.matrix.r
\docType{methods}
\name{show,snps.matrix-method}
\alias{show,snps.matrix-method}
\title{Show method of snps.matrix.}
\usage{
\S4method{show}{snps.matrix}(object)
}
\arguments{
\item{object}{an snps.matrix object}
}
\description{
Show method of snps.matrix.
}
