% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.segments.r
\docType{methods}
\name{show,snps.segments-method}
\alias{show,snps.segments-method}
\title{Show method of snps.segments.}
\usage{
\S4method{show}{snps.segments}(object)
}
\arguments{
\item{object}{the snps.segments object}
}
\description{
Show method of snps.segments.
}
