% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsMatrix.r
\name{show,snpsMatrix-method}
\alias{show,snpsMatrix-method}
\title{Show method of snpsMatrix.}
\usage{
\S4method{show}{snpsMatrix}(object)
}
\arguments{
\item{object}{an snpsMatrix object}
}
\description{
Show method of snpsMatrix.
}
