% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsSegments.r
\docType{methods}
\name{show,snpsSegments-method}
\alias{show,snpsSegments-method}
\title{Show method of snpsSegments.}
\usage{
\S4method{show}{snpsSegments}(object)
}
\arguments{
\item{object}{the snpsSegments object}
}
\description{
Show method of snpsSegments.
}
