% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submapsList.r
\docType{methods}
\name{show,submapsList-method}
\alias{show,submapsList-method}
\title{Show method of submapsList.}
\usage{
\S4method{show}{submapsList}(object)
}
\arguments{
\item{object}{an submapsList object}
}
\description{
Show method of submapsList.
}
