% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.matrix.r
\docType{class}
\name{snps.matrix-class}
\alias{snps.matrix-class}
\title{Class snps.matrix}
\description{
Class \code{snps.matrix} This class is use to create an object that will represent the submap created
}
\section{Slots}{

\describe{
\item{\code{step}}{the value of the step used to created the segments}

\item{\code{submap}}{a vector of index}
}}

