% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.segments.r
\docType{class}
\name{snps.segments-class}
\alias{snps.segments-class}
\title{Class snps.segments.}
\description{
Class \code{snps.segments} defines snps.segments object that will store the list of segments created.
}
\section{Slots}{

\describe{
\item{\code{gap}}{the value of the gap used to created the segments}

\item{\code{unit}}{the unit of the marker in the segments (cM or bP)}

\item{\code{snps.segments}}{the list of segments}
}}

