% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsMatrix.r
\docType{class}
\name{snpsMatrix-class}
\alias{snpsMatrix-class}
\title{Class snpsMatrix}
\description{
Class \code{snpsMatrix} This class is use to create an object that will represent the submap created
}
\section{Slots}{

\describe{
\item{\code{step}}{the value of the step used to created the segments}

\item{\code{submap}}{a vector of index}
}}

