% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpsSegments.r
\docType{class}
\name{snpsSegments-class}
\alias{snpsSegments-class}
\title{Class snpsSegments.}
\description{
Class \code{snpsSegments} defines snpsSegments object that will store the list of segments created.
}
\section{Slots}{

\describe{
\item{\code{gap}}{the value of the gap used to created the segments}

\item{\code{unit}}{the unit of the marker in the segments (cM or bP)}

\item{\code{snpsSegments}}{the list of segments}
}}

