% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submapEstim.r
\name{submapEstim}
\alias{submapEstim}
\title{Summury of estimation through the submaps}
\usage{
submapEstim(submaps)
}
\arguments{
\item{submaps}{a list of submaps}
}
\value{
this fonction returns a list of dataframe.
}
\description{
This function creates a summary of the a and f estimationq on submaps created.
}
\details{
The first element of the returned list is the estimation of f through the submaps.

The second element of the returned list is the estimation of a through the submaps.
}
\examples{
 

}
\seealso{
\code{\link{}}
}
