% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submapLikelihood.r
\name{submapLikelihood}
\alias{submapLikelihood}
\title{Estimation of likelihood0 and likelihood2}
\usage{
submapLikelihood(submaps)
}
\arguments{
\item{submaps}{a list of submaps}
}
\value{
this function returns a list of dataframe.
}
\description{
This function creates a summary on the submaps created.
}
\details{
the first element of the list is the estimation of likelihood0

the second element of the list is the estimation of likelihood1
}
\examples{
 

}
\seealso{
setSummary
}
