% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submapSummary.r
\name{submapSummary}
\alias{submapSummary}
\title{Submap summary}
\usage{
submapSummary(submaps, a.threshold = 1)
}
\arguments{
\item{submaps}{a list of submaps}

\item{a.threshold}{the maximum value for a (default is 1)}
}
\value{
this function returns a dataframe.
}
\description{
This function creates a summary on the submaps created.
}
\details{
This function gives for each genotyped individual summary statistics about the calculations.

This function returns a dataframe with 13 columns :

- FID : family identifier

- IID : individual identifier

- STATUS : status (1 non affected, 2 affected, 0 unknown)

- submaps : number of submaps used

- QUALITY: percentage of valid submaps (i.e. submaps with a < 1)

- F_MIN: minimum f on valid submaps

- F_MAX: maximum f on valid submaps

- F_MEAN: mean f on valid submaps

- F_MEDIAN: median f on valid submaps (recommended to estimate f)

- A_MEDIAN: median a on valid submaps (recommended to estimate a)

- pLRT_MEDIAN: median p-value of LRT tests on valid submaps

- INBRED: a flag indicating if the individual is inbred (pLRT_MEDIAN <0.05) or not

- pLRT_<0.05: number of valid submaps with a LRT having a p-value below 0.05
}
\seealso{
setSummary
}
