% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMap.r
\name{summaryMap}
\alias{summaryMap}
\title{Markers selected through the submaps}
\usage{
summaryMap(atlas)
}
\arguments{
\item{submaps}{a list containing the different submaps created}
}
\value{
This function returns a dataframe.
}
\description{
This function  will give the number of markers selected in your submaps.
}
\details{
make sure to provide a list of submaps in argument. This function returns a dataframe with columns :

\describe{
 \item{id}{id of the snps}
 \item{chr}{chromosome}
 \item{pos}{position in bp}
 \item{dist}{position in cM}
 \item{Freq}{number of times the marker has been picked}
}
}
\seealso{
setSummary
}
