% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMarker.r
\name{summaryMarker}
\alias{summaryMarker}
\title{How many markers have been selected 0, 1, ... times}
\usage{
summaryMarker(atlas)
}
\arguments{
\item{atlas}{an atlas}
}
\value{
A data frame with columns
\describe{
  \item{markers}{number of markers}
  \item{picked}{number of times picked}
}
}
\description{
Gives the number of markers of the initial bedmatrix that have never been picked in any submap,
that have been picked once, etc.
}
\seealso{
setSummary
}
