% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubmapBySnps.r
\name{createSubmapBySnps}
\alias{createSubmapBySnps}
\title{Creation of a submaps}
\usage{
createSubmapBySnps(bedmatrix, segmentsList, epsilon = 0.001, fileName)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{segmentsList}{a list of segment for each chromosomes}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{fileName}{You have the possibility to pass your own list of marker, that is to say, your own submaps}
}
\value{
return an snpsMatrix object with some of his slots filled.
}
\description{
This function creates a submaps using the list segments created by using gap between markers
}
\details{
This function will iterates over the list of segments, then for each segments it will pick randomly one marker 
and from this marker chose a marker every "step" step from back to end.

Once the iteration over the list is over, the function will create an object and filled some of his slot.

If you are using a fileName, please make sure to have one marker per line
}
\examples{
 
#Please refer to vignette 


}
\seealso{
makeSubmapsBySnps
}
