% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fantasio.r, R/data.r
\docType{package}
\name{Fantasio}
\alias{Fantasio}
\alias{Fantasio}
\alias{Fantasio-package}
\title{Wrapper for the package Fantasio}
\usage{
Fantasio(bedmatrix, segments = c("Hotspots", "Distance"), segment.options,
  n = 100, n.cores = 1, epsilon = 0.001, run.festim = TRUE, list.id,
  run.proba = TRUE, recap.by.segments = FALSE, verbose = TRUE,
  debug = FALSE, threshold = 0.5, q = 1e-04, quality = 95,
  n.consecutive.marker = 5)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{segments}{The method to use for submap creation ("Hotspots" or "Distance")}

\item{segment.options}{a  list of arguments to the function that will create the segments list. The names attribute of args gives the argument names.}

\item{n}{the number of submaps wanted(default is 100)}

\item{n.cores}{the number of cores to use if you want to compute submaps using parellelism (default is 1)}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{run.festim}{whether you want to computes a, f, p.lrt, likelihood0/1 for each submaps (default is TRUE)}

\item{list.id}{a list of individuals from which to computes HBD, 
FLOD score and HFLOD score (see details for more informations)}

\item{run.proba}{whether you want to computes HBD, FLOD score and HFLOD score (default is TRUE)}

\item{recap.by.segments}{if you want the summary of probabilities by snps or by segments (default is FALSE)}

\item{verbose}{whether you want informations about computations (default is TRUE)}

\item{debug}{whether you want advanced output for the computation process (default is FALSE)}

\item{threshold}{the value of the threshold when finding HBD segment, threshold is the probability of being HBD or not (default is 0.5)}

\item{q}{Allows the user to choose the assumed frequency of the mutation involved in the disease for each individual (default is 0.0001)}

\item{quality}{Allows the user to choose the minimal quality (in \%) to include an inbred individual into the analysis (default is 95)}

\item{n.consecutive.marker}{the number of consecutive marker with a probabilitie equal or greater to the value of threshold, to be use to fing HBD segments}
}
\description{
This function is used as a wrapper for the package Fantasio to create segments list and submaps

Fantasio.
}
\details{
This function is a wrapper to make the usage of Fantasio, the package, easier. The function calls differents function :

The first function `createSegmentsListBySnps` and`createSegmentsListByHotspots` is used to create a list of segments though the genome.

The second function `makeAllSubmapsBySnsp`and `makeAllSubmapsByHotspots` is used to create submaps.

The segments arguments accept only two options : Hotspots or Distance.

Depending of the value of the segments argument (Hotspots or Distance) the segments are made by using the hotspots of recombinaison

in the genome thanks to hotspots file, then the submaps are made by Hotspots by picking randomly a maker in every segments created before.

Or the segments are made by using the gaps between markers (option Distance), then the submaps are made by picking a random marker in

every segments and going through each segment from left to right using a supply step (by default it is 0.5 cM).

To see which arguments can be passed in the list of args in the segment.options please go check the createSegmentListBySnps/Hotspots functions manual.

When doing a number N of submaps using the options recap.by.segments,

the different values computed from the

snps picked randomly (a, f, p.lrt, ...) in a segments is considered different values

for the same segments.
}
\examples{
##install.packages("HGDP.CEPH", repos="https://genostats.github.io/R/") ## make this only one time
require(Fantasio)
require(HGDP.CEPH)
filepath <-system.file("extdata", "hgdp_ceph.bed", package="HGDP.CEPH")
x <- read.bed.matrix(filepath)
x <- set.stats(x)
x.me <- select.inds(x, population == "Bedouin")
x.me@ped$pheno <- rep(2,48) #The package only analyzes individualw with a status of 2
submaps <- Fantasio(x.me, "Hotspots", n=5)

}
\seealso{
read.bed.matrix

createSegmentsListBySnps

makeAllSubmapsBySnps

makeAllSubmapsByHotspots

makeAllSubmapsByHotspots
}
