% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/festim.r
\name{festim}
\alias{festim}
\title{Computation of a and f}
\usage{
festim(x, verbose = TRUE, debug = FALSE)
}
\arguments{
\item{x}{a submap matrix}

\item{verbose}{whether you want informations about the computation process (default is TRUE)}

\item{debug}{whether you want advanced output about the computation process (default is FALSE)}

\item{segmentsList}{a list of segments for each chromosome}
}
\value{
returns the submap matrix with the slot a, f, p.lrt, likelihood0 and likelihood1 filled.
}
\description{
This function is used to compute a and f and other statistical values (p.lrt, likelihood0, likelihood1)
}
\details{
This function is using the optim function with an L-BGFS method to find the best value for a and f
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListBySnps(bedMatrix)
submaps <- makeSubmapsBySnps(bedMatrix, 5, segmentList) #<- inside this function is called the festim function

}
