% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmHBDPlot.r
\name{glmHBDPlot}
\alias{glmHBDPlot}
\title{QQ-Plot & Manhattan Plots for glm on HBD prob or FLOD}
\usage{
glmHBDPlot(x, expl_var, save = FALSE)
}
\arguments{
\item{x}{an atlas object}

\item{expl_var}{the explanatory variable 'FLOD' or 'HBD_prob'}

\item{save}{choose if plot are saved or not (.png format) (default = FALSE)}

\item{n.cores}{number of cores for parallelization calculation (default = 1)}
}
\description{
QQ-Plot & Manhattan Plots for glm on HBD prob or FLOD
}
\seealso{
\code{\link{glmHBD}}
}
