% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBD.plot.chr.r
\name{HBD.plot.chr}
\alias{HBD.plot.chr}
\title{plot of HBD segment}
\usage{
HBD.plot.chr(Submaps, ROHfile, unit = "cM", chr, list.ids, regions, outfile)
}
\arguments{
\item{Submaps}{a list.submap object}

\item{ROHfile}{a ROH file from which the segments will be plotted (optional)}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{chr}{the chromosome number from which to plot HBD segment}

\item{list.ids}{a vector containing a list of individuals from which only the HBD segments for this chromosome will be ploted (optional)}

\item{regions}{a specific region to be enlighted in the plot (optional)}

\item{outfile}{a name for the plot (optional)}
}
\value{
return a plot of the chromosome HBD segments for all the individual
}
\description{
This function plots the HBD segments for a given chromosom and all the individual
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListBySnps(bedMatrix)
submaps <- makeSubmapsBySnps(bedMatrix, 5, segmentList)
HBD.plot.chr(submaps, chr=1)

}
