% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBD.plot.id.r
\name{HBD.plot.id}
\alias{HBD.plot.id}
\title{plot of HBD segment}
\usage{
HBD.plot.id(Submaps, ROHfile, unit = "cM", individual.id, family.id, regions,
  outfile, build = 37)
}
\arguments{
\item{Submaps}{a list.submap object}

\item{ROHfile}{a ROH file from which the segments will be plotted (optional)}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{individual.id}{the individual id of the individual wanted}

\item{family.id}{the family id of the individual wanted}

\item{regions}{a specific region to be enlighted in the plot (optional)}

\item{outfile}{a name for the plot (optional)}

\item{build}{the value of the build to use to plot chromosome in the plot value accepted are 35, 36, 37, 38 (default is 37)}
}
\value{
return a plot of the individual's HBD segments.
}
\description{
This function plots the HBD segments for all the chromosoms of a given individual
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
\examples{
 
##install.packages("HGDP.CEPH", repos="https://genostats.github.io/R/") ## make this only one time
require(Fantasio)
require(HGDP.CEPH)
filepath <-system.file("extdata", "hgdp_ceph.bed", package="HGDP.CEPH")
x <- read.bed.matrix(filepath)
x <- set.stats(x)
x.me <- select.inds(x, population == "Bedouin")
x.me@ped$pheno <- rep(2,48) #The package analyzes only individualw with a status of 2
submaps <- Fantasio(x.me, "Hotspots", n=5)
HBD.plot.id(submaps, individual.id="HGDP00623", family.id="HGDP00623")

}
