% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFLOD.manhattan.plot.r
\name{HFLOD.manhattan.plot}
\alias{HFLOD.manhattan.plot}
\title{Creation of an manhattan plot of the HFLOD}
\usage{
HFLOD.manhattan.plot(submaps, regions, unit = "cM")
}
\arguments{
\item{submaps}{a list.submaps object}

\item{regions}{a matrix containing the value to be highlighted in the plot}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}
}
\value{
This function returns a manhattan plot of all the HFLOD score over all the chromosome
}
\description{
This fonction to plot a manhanttan plot of the HFLOD score
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

- the chromosome number

- start

- end
}
\examples{
##install.packages("HGDP.CEPH", repos="https://genostats.github.io/R/") ## make this only one time
require(Fantasio)
require(HGDP.CEPH)
filepath <-system.file("extdata", "hgdp_ceph.bed", package="HGDP.CEPH")
x <- read.bed.matrix(filepath)
x <- set.stats(x)
x.me <- select.inds(x, population == "Bedouin")
x.me@ped$pheno <- rep(2,48) #The package analyzes only individualw with a status of 2
submaps <- Fantasio(x.me, "Hotspots", n=5)
HFLOD.manhattan.plot(submaps)
}
\seealso{
set.HFLOD
}
