% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFLOD.plot.submaps.r
\name{HFLOD.plot.chr}
\alias{HFLOD.plot.chr}
\title{Plot of the HFLOD}
\usage{
HFLOD.plot.chr(submaps, unit = "cM", chr, regions, color2 = "green4",
  nbSNP_MA = 50)
}
\arguments{
\item{submaps}{a list.submaps object}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{chr}{the chromosome number from which to plot HFLOD score}

\item{regions}{a matrix containing the value to ve highlighted in the plot}

\item{color2}{the color of the regions highlighted (default is "green4")}

\item{nbSNP_MA}{number of SNP for the moving average (default is 50)}
}
\value{
This function returns a manhattan plot of all the HFLOD score over all the chromosome
}
\description{
This fonction plot the HFLOD score for a chromosome
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

- the chromosome number

- start

- end
}
\examples{
 
##install.packages("HGDP.CEPH", repos="https://genostats.github.io/R/") ## make this only one time
require(Fantasio)
require(HGDP.CEPH)
filepath <-system.file("extdata", "hgdp_ceph.bed", package="HGDP.CEPH")
x <- read.bed.matrix(filepath)
x <- set.stats(x)
x.me <- select.inds(x, population == "Bedouin")
x.me@ped$pheno <- rep(2,48) #The package analyzes only individualw with a status of 2
submaps <- Fantasio(x.me, "Hotspots", n=5)
HFLOD.manhattan.plot(submaps)
}
\seealso{
set.HFLOD
}
