% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hotspot}
\alias{hotspot}
\alias{hotspot_hg17}
\alias{hotspot_hg18}
\alias{hotspot_hg19}
\title{Position of hotspots throughout the human genome}
\format{Data frames with 32996 rows and 7 variables}
\usage{
data(hotspot_hg17)
}
\description{
These datasets are data frames containing position and intensity of hotspots in the human genome,
for releases hg17 (build 35), hg18 (build 36) and hg19 (build 37).
}
\details{
The variables are:
\itemize{
  \item Chromosome. Chromosome number
  \item Centre. Center of the hotspots in Mb
  \item Start. Start of the hotspot
  \item End.  End of the hotspot
  \item Widthkb. Width of the hotposts in kb
  \item IntensitycMMb. Intensity in M / Mb
  \item TotaldistancecM. Total distance in cM
}
}
\keyword{datasets}
