% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.plot.chr.r
\name{ROH.plot.chr}
\alias{ROH.plot.chr}
\title{Creation of plot using ROH file}
\usage{
ROH.plot.chr(ROHfile, submaps, unit = "cM", chr, outfile, listid, regions)
}
\arguments{
\item{ROHfile}{an ROH file}

\item{submaps}{a list.submaps object}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{chr}{the chromosome number from which to plot ROH}

\item{outfile}{the name of the plot}

\item{listid}{a vector containing the family id follow by the individual id}

\item{regions}{a matrix containing the value to ve highlighted in the plot}
}
\value{
This function returns a plot of the HBD segments for a specific chromosome
}
\description{
This function is use to plot HBD segments for a chromosome using an ROH file
}
\details{
Use this function when you want to use an ROH file to plot your HBD segments for a specific chromosome.

Two unit are accepted : "Bases" or "cM".

If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListByHotspots(bedMatrix)
submaps <- makeSubmapsByHotspots(bedMatrix, 10, segmentList)  
ROH.plot.chr(yourROHfile, submaps)
}
