% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.plot.id.r
\name{ROH.plot.id}
\alias{ROH.plot.id}
\title{Creation of plot using ROH file}
\usage{
ROH.plot.id(ROHfile, submaps, unit = "cM", regions, outfile, family_id,
  individual_id, save_file = F, build = 37)
}
\arguments{
\item{ROHfile}{an ROH file}

\item{submaps}{a list.submaps object}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "CM")}

\item{regions}{a matrix containing the value to ve highlighted in the plot}

\item{outfile}{the name of the plot}

\item{family_id}{the individual id of the individual wanted}

\item{individual_id}{the family id of the individual wanted}

\item{save_file}{whether you want or not to save the plot}
}
\value{
This function returns a plot of the HBD segments for a specific individual
}
\description{
This function is use to plot HBD segments for a specific infdividual using an ROH file
}
\details{
Use this function when you want to use an ROH file to plot your HBD segments for a specific individual.

Two unit are accepted : "Bases" or "cM".

If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

-the chromosome number

-start

-end
}
