% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsListSummary.r
\name{segmentsListSummary}
\alias{segmentsListSummary}
\title{Summary of marker picked}
\usage{
segmentsListSummary(segmentList)
}
\description{
This function is uses to return the number of segments and number of marker in the list of segments outputed by the function 
`createSegmentsListByHotspots`
}
