% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsListSummary.r
\name{segmentsListSummary}
\alias{segmentsListSummary}
\title{Summary of marker picked}
\usage{
segmentsListSummary(segmentList)
}
\arguments{
\item{segmentList}{A list of segments}
}
\description{
This function is uses to return the number of segments and number of marker in the list of segments outputed by the function 
`createSegmentsListByHotspots`
}
\examples{
##install.packages("HGDP.CEPH", repos="https://genostats.github.io/R/") ## make this only one time
require(Fantasio)
require(HGDP.CEPH)
filepath <-system.file("extdata", "hgdp_ceph.bed", package="HGDP.CEPH")
x <- read.bed.matrix(filepath)
x <- set.stats(x)
x.me <- select.inds(x, population == "Bedouin")
x.me@ped$pheno <- rep(2,48) #The package analyzes only individualw with a status of 2
s <- createSegmentsListByHotspots(x.me)
segmentsListSummary(s)

}
