% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.HFLOD.r
\name{set.HFLOD}
\alias{set.HFLOD}
\title{Computation of HFLOD scores}
\usage{
set.HFLOD(submaps)
}
\arguments{
\item{submaps}{a list.submaps object}
}
\value{
the list.submaps object with it's slot HFLOD filled with a matrix of dimensions : number_of_marker x 2. The first column is the value of HFLOD for the marker. 
The second value is the moving average probability.
}
\description{
This function is used to compute HFLOD scores on individuals in a sample
}
\details{
This function iterates over the slots atlas of the list.submaps object.
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListByHotspots(bedMatrix)
individualList <- c("familyID0_individualID0", "familyID1_individualID2"), "familyID2_individualID2")
makeSubmapsByHotspots(bedMatrix, 10, segmentList, list.id=individualList)  #the function set.HFLOD is use inside the function setSummary of this function
}
\seealso{
\code{\link{set.HBD.prob}}

\code{\link{set.FLOD}}
}
