% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFLOD.r
\name{setFLOD}
\alias{setFLOD}
\title{Computation of FLOD scores}
\usage{
setFLOD(submaps, w.id, q = 1e-04)
}
\arguments{
\item{submaps}{a atlas object}

\item{w.id}{The indices of individals for which FLOD are computed}

\item{q}{The assumed frequency of the mutation involved in the disease for each individual. Default is 0.0001.}
}
\value{
the atlas object with each FLOD.prob slot of each submaps in the slot submaps_list computed
}
\description{
This function is uses to compute FLOD scores on individual present in a sample
}
\details{
This function iterates over the slots submaps_list of the atlas object.

For each submaps in the slots submaps_list of the object, the slot FLOD.prob will be filled with a matrix of dimension : number_inidividual x number_of_markers
}
\seealso{
setHBDProb

setHFLOD
}
