% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.summary.r
\name{setSummary}
\alias{setSummary}
\title{Creation summary statistic files}
\usage{
setSummary(submaps, list.id, run_a_f = TRUE, probs = TRUE,
  by_segments = FALSE, q = 1e-04, threshold = 0.5, quality = 95,
  n.consecutive.marker = 5)
}
\arguments{
\item{submaps}{a list.submaps object}

\item{list.id}{you can either :
- ignore this parameter if you want to compute HBD, FLOD and HFLOD 
  for individuals who are considerated INBRED and with a QUALITY
  greater or equal to 95%}
- enter a list of individual for a computation of HBD, FLOD score HFLOD score for them
- use "all" for a computation of HBD, FLOD score and HFLOD score for every individual}

\item{run_a_f}{a flag indicating if the estimation of a and f has been computed for the submaps using `festim` (default is TRUE)}

\item{probs}{a flag indicating if the HBD probabilities and FLOD score has been computed for the submaps (default is TRUE)}

\item{by_segments}{if the summary files has to be computed considering segments or snps (defaut is FALSE) 
(more information in the documentation of `makeSubmapsByHotspots` or`makeSubmapsBySnps` functions)}

\item{q}{Allows the user to choose the assumed frequency of the mutation involved in the disease for each individual (default is 0.0001)}

\item{threshold}{the value of the threshold when finding HBD segment, threshold is the probability of being HBD or not (default is 0.5)}

\item{quality}{The minimum percentage use to assume if a submap is valid (default is 95)}

\item{n.consecutive.marker}{the number of consecutive marker with a probabilitie equal or greater to the value of the threshold, to be use to fing HBD segments (default is 5)}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{run.festim}{whether you want to computes a, f, p.lrt, likelihood0/1 for each submaps}

\item{run.proba}{whether you want to computes HBD, FLOD score and HFLOD score}

\item{test}{a vector containing the index of the different individual with a STATUS of 2 (individual with the disease)}
}
\value{
return a new list object containing every dataframe and object created
}
\description{
This function is uses to ouput all the summary files after the creation of the submaps by the following two functions : 
`makeSubmapsByHotspots` and `makeSubmapsBySnps` is over.
}
\details{
the function filled the empty slots of the list.submaps object. It calls different functions and uses the results of each one to filled 
the object. This function computes summary for the following elements :

- summary for likelihood0/likelihood1

- summary for estimation of a and f

- marker summary

- summary for the submaps

- summary for HBD probabilities

- summary for HFLOD

- comutation of HBD segments
}
\examples{
 
bedMatrix <- read.bed.matrix("yourFile")
segmentList <- createSegmentsListByHotspots(bedMatrix)
submaps <- makeSubmapsByHotspots(bedMatrix, 5, segmentList) #this function is a wrapper that uses the function setSummary

}
\seealso{
\code{\link{makeSubmapsByHotspots}}

\code{\link{createSegmentsListByHotspots}}

\code{\link{festim}}

\code{\link{set.HBD.prob}}

\code{\link{set.FLOD}}

\code{\link{set.HFLOD}}

\code{\link{recap}}

\code{\link{setSummary}}

\code{\link{submapLikelihood}}

\code{\link{submapEstim}}

\code{\link{summaryMarker}}

\code{\link{submapSummary}}

\code{\link{HBD.segments}}
}
