% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMap.r
\name{summaryMap}
\alias{summaryMap}
\title{Number of selected marker through the submaps}
\usage{
summaryMap(submaps)
}
\arguments{
\item{submaps}{a list containing the different submaps created}
}
\value{
This function returns a dataframe.
}
\description{
This function  will give the number of markers selected in your submaps.
}
\details{
make sure to provide a list of submaps in argument. This function returns a dataframe with 2 columns :

-snsp : the id of the snps

-Freq : the number of times the marke has been picked
}
\seealso{
setSummary
}
