% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMarker.r
\name{summaryMarker}
\alias{summaryMarker}
\title{Number of marker selected in a submap}
\usage{
summaryMarker(submaps, bedmatrix)
}
\arguments{
\item{submaps}{a list of submaps}

\item{bedmatrix}{a bed.matrix object}
}
\value{
this function returns a dataframe.
}
\description{
This function gives for each marker the number of time that it has been selected in a submap.
}
\details{
the dataframe contains the following elements :

- by columns is the number of markers used

- by row is the number of time picked
}
\examples{
 

}
\seealso{
setSummary
}
